requirements/base.txt
celery==5.2.7

Django==3.2.18
dj-database-url==1.2.0
django-model-utils==4.3.1
django-cors-headers==3.13.0
django-avatar==7.1.1
django-jinja==2.10.2
django-filter==2.4.0
django-crispy-forms==1.14.0
easy-thumbnails==2.8.5
django-fsm==2.8.1
django-fsm-admin==1.2.5
django-phonenumber-field==7.0.2
django-constance[database]==2.9.1

django-picklefield==3.0.1

# sending emails
djmail==2.0.0

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-expander==0.2.3
drf-extra-fields==3.4.1
rest-social-auth==5.0.1

# postgresql driver
psycopg2==2.9.5

# runs the app server
gunicorn==20.1.0
daphne==4.0.0
uvicorn[standard]==0.20.0

# unique id generator
hashids==1.3.1

# datetime helpers
python-dateutil==2.8.2

# phone number parser
phonenumbers==8.13.4

# JSON log formatter
JSON-log-formatter==0.3.1

#redis
redis==4.5.5
django_redis==4.11.0

# channels
channels==4.0.0
channels-redis==4.1.0
djangochannelsrestframework==0.3.0

# vector search
pgvector==0.1.6
numpy==1.23.2
tiktoken==0.3.2
sentence-transformers==2.2.2
transformers==4.28.1

# HTML
beautifulsoup4==4.12.2

# AI
openai==0.27.8
langchain==0.0.198
